+++ scienscope_2021/ProjectSettings/EditorBuildSettings.asset
@@ -5,7 +5,7 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/0_start.unity
     guid: 0f581a67f8826f34f87dc61471815a02
   - enabled: 1
