+++ scienscope_2021/UserSettings/EditorUserSettings.asset
@@ -27,10 +27,10 @@ EditorUserSettings:
       value: 224247031146464f08031c2e11375a2f05141d242966333e243d04
       flags: 0
     RecentlyUsedScenePath-7:
-      value: 22424703114646680e0b0227036c0e2f2431310e133f273939201334acf53a31f6fe
+      value: 22424703114646680e0b0227036c0c2f0612173a202d68252320092a
       flags: 0
     RecentlyUsedScenePath-8:
-      value: 22424703114646680e0b0227036c0c2f0612173a202d68252320092a
+      value: 22424703114646680e0b0227036c0e2f2431310e133f273939201334acf53a31f6fe
       flags: 0
     RecentlyUsedScenePath-9:
       value: 22424703114646680e0b0227036c0d2f140519232266333e243d04
